refactor(session): extract audit-trail methods to audit_store.go
AuditRecord, ErrAuditChainInvalid, AppendAudit, and GetAuditTrail moved
from store.go to audit_store.go. Reduces store.go by ~100 lines and
isolates audit concern. Methods remain on *SQLiteStore (same package).

Ref #69
